% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_time_to_event.R
\name{simulate_time_to_event}
\alias{simulate_time_to_event}
\title{Simulate heterogeneous time to event data}
\usage{
simulate_time_to_event(n, mean_death_other_cause = 1, mean_death_cancer = 1)
}
\arguments{
\item{n}{sample size to simulate}

\item{mean}{population mean}
}
\description{
Simulate heterogeneous time to event data
}
