% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{Vickers-Life-Expectancy-Model-Bias-package}
\alias{Vickers-Life-Expectancy-Model-Bias}
\alias{Vickers-Life-Expectancy-Model-Bias-package}
\title{Vickers-Life-Expectancy-Model-Bias: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
